+++ instant_pot/src/main/java/com/web_services/instant_pot/domain/review/ReviewLogic.java
@@ -8,27 +8,28 @@ import antlr.collections.List;
 
 public class ReviewLogic {
 	
-	public Review getReviewByID(int reviewID) {
+	public Review getReviewByID(long reviewID) {
 		ReviewDAL rd = new ReviewDAL();
 		return rd.getReviewByID(reviewID);
 	}
 	
-	public HashSet<Review> getAllReviewForCustomer(int custID){
+	public HashSet<Review> getAllReviewForCustomer(long custID){
 		ReviewDAL rd = new ReviewDAL();
 		return rd.getAllReviewForCustomer(custID);
 	}
 
-	public HashSet<Review> getAllReviewForProduct(int productID){
+	public HashSet<Review> getAllReviewForProduct(long productID){
 		ReviewDAL rd = new ReviewDAL();
 		return rd.getAllReviewForProduct(productID);
 	}
 	
 	public Review addReview(Long custID, Long productID, String reviewBody) {
+
 		ReviewDAL rd = new ReviewDAL();
 		return rd.addReview(custID, productID, reviewBody);
 	}
 	
-	public static Review deleteReview(int reviewID) {
+	public static Review deleteReview(long reviewID) {
 		ReviewDAL rd = new ReviewDAL();
 		return rd.deleteReview(reviewID);
 	}
